recommit reverted changes (refactoring and test suit)
