+++ ledsa/tests/AcceptanceTests/LedsaATestLibrary.py
@@ -1,0 +1,87 @@
+from robot.api.deco import keyword, library
+from robot.libraries.BuiltIn import BuiltIn
+import os
+from PIL import Image
+import numpy as np
+from scipy.stats import norm
+from ledsa.core.ledsa_conf import ConfigData
+from subprocess import Popen, PIPE
+
+
+@library
+class LedsaATestLibrary:
+
+    @keyword
+    def change_dir(self, new_dir):
+        os.chdir(new_dir)
+
+    @keyword
+    def create_test_image(self, amount=1):
+        img_array = create_img_array()
+        for i in range(amount):
+            img = Image.fromarray(img_array, 'RGB')
+            img.save(f'test_img_{i}.png', exif=b'Exif\x00\x00II*\x00\x08\x00\x00\x00\x0c\x00\x0f\x01\x02\x00\x06\x00\x00\x00\x9e\x00\x00\x00\x10\x01\x02\x00\x0e\x00\x00\x00\xa4\x00\x00\x00\x12\x01\x03\x00\x01\x00\x00\x00\x01\x00\x00\x00\x1a\x01\x05\x00\x01\x00\x00\x00\xc4\x00\x00\x00\x1b\x01\x05\x00\x01\x00\x00\x00\xcc\x00\x00\x00(\x01\x03\x00\x01\x00\x00\x00\x02\x00\x00\x002\x01\x02\x00\x14\x00\x00\x00\xd4\x00\x00\x00;\x01\x02\x00\x01\x00\x00\x00\x00\x00\x00\x00\x13\x02\x03\x00\x01\x00\x00\x00\x02\x00\x00\x00\x98\x82\x02\x00\x01\x00\x00\x00\x00\x00\x00\x00i\x87\x04\x00\x01\x00\x00\x00h\x01\x00\x00%\x88\x04\x00\x01\x00\x00\x00\x84)\x00\x00\x860\x00\x00Canon\x00Canon EOS 80D\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00H\x00\x00\x00\x01\x00\x00\x00H\x00\x00\x00\x01\x00\x00\x002018:11:27 15:46:03\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00&\x00\x9a\x82\x05\x00\x01\x00\x00\x006\x03\x00\x00\x9d\x82\x05\x00\x01\x00\x00\x00>\x03\x00\x00"\x88\x03\x00\x01\x00\x00\x00\x01\x00\x00\x00\'\x88\x03\x00\x01\x00\x00\x00d\x00\x00\x000\x88\x03\x00\x01\x00\x00\x00\x02\x00\x00\x002\x88\x04\x00\x01\x00\x00\x00d\x00\x00\x00\x00\x90\x07\x00\x04\x00\x00\x000230\x03\x90\x02\x00\x14\x00\x00\x00F\x03\x00\x00\x04\x90\x02\x00\x14\x00\x00\x00Z\x03\x00\x00\x01\x91\x07\x00\x04\x00\x00\x00\x01\x02\x03\x00\x01\x92\n\x00\x01\x00\x00\x00n\x03\x00\x00\x02\x92\x05\x00\x01\x00\x00\x00v\x03\x00\x00\x04\x92\n\x00\x01\x00\x00\x00~\x03\x00\x00\x07\x92\x03\x00\x01\x00\x00\x00\x05\x00\x00\x00\t\x92\x03\x00\x01\x00\x00\x00\x10\x00\x00\x00\n\x92\x05\x00\x01\x00\x00\x00\x86\x03\x00\x00|\x92\x07\x00\n$\x00\x00\x8e\x03\x00\x00\x86\x92\x07\x00\x08\x01\x00\x00\x98\'\x00\x00\x90\x92\x02\x00\x03\x00\x00\x0003\x00\x00\x91\x92\x02\x00\x03\x00\x00\x0003\x00\x00\x92\x92\x02\x00\x03\x00\x00\x0003\x00\x00\x00\xa0\x07\x00\x04\x00\x00\x000100\x01\xa0\x03\x00\x01\x00\x00\x00\x01\x00\x00\x00\x02\xa0\x03\x00\x01\x00\x00\x00p\x17\x00\x00\x03\xa0\x03\x00\x01\x00\x00\x00\xa0\x0f\x00\x00\x05\xa0\x04\x00\x01\x00\x00\x00\xa0(\x00\x00\x0e\xa2\x05\x00\x01\x00\x00\x00\xbe(\x00\x00\x0f\xa2\x05\x00\x01\x00\x00\x00\xc6(\x00\x00\x10\xa2\x03\x00\x01\x00\x00\x00\x02\x00\x00\x00\x01\xa4\x03\x00\x01\x00\x00\x00\x00\x00\x00\x00\x02\xa4\x03\x00\x01\x00\x00\x00\x01\x00\x00\x00\x03\xa4\x03\x00\x01\x00\x00\x00\x01\x00\x00\x00\x06\xa4\x03\x00\x01\x00\x00\x00\x00\x00\x00\x000\xa4\x02\x00\x01\x00\x00\x00\x00\x00\x00\x001\xa4\x02\x00\r\x00\x00\x00\xee(\x00\x002\xa4\x05\x00\x04\x00\x00\x00\x0e)\x00\x004\xa4\x02\x00\x08\x00\x00\x00.)\x00\x005\xa4\x02\x00\x0b\x00\x00\x00x)\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\xf4\x01\x00\x00\x10\x00\x00\x00\x01\x00\x00\x002018:11:27 15:46:03\x002018:11:27 15:46:03\x00\x00\x00\t\x00\x00\x00\x01\x00\x00\x00\x08\x00\x00\x00\x01\x00\x00\x00\x00\x00\x01\x00\x00\x00\x12\x00\x00\x00\x01\x00\x00\x00*\x00\x01\x00\x03\x001\x00\x00\x00\x8c\x05\x00\x00\x02\x00\x03\x00\x04\x00\x00\x00\xee\x05\x00\x00\x03\x00\x03\x00\x04\x00\x00\x00\xf6\x05\x00\x00\x04\x00\x03\x00"\x00\x00\x00\xfe\x05\x00\x00\x06\x00\x02\x00\x0e\x00\x00\x00B\x06\x00\x00\x07\x00\x02\x00\x18\x00\x00\x00b\x06\x00\x00\t\x00\x02\x00 \x00\x00\x00z\x06\x00\x00\r\x00\x07\x00\x00\x06\x00\x00\x9a\x06\x00\x00\x10\x00\x04\x00\x01\x00\x00\x00P\x03\x00\x80\x13\x00\x03\x00\x04\x00\x00\x00\x9a\x0c\x00\x00\x19\x00\x03\x00\x01\x00\x00\x00\x01\x00\x00\x00&\x00\x03\x00\xc6\x00\x00\x00\xa2\x0c\x00\x005\x00\x04\x00\x04\x00\x00\x00.\x0e\x00\x008\x00\x07\x00L\x00\x00\x00>\x0e\x00\x00\x93\x00\x03\x00&\x00\x00\x00\x8a\x0e\x00\x00\x95\x00\x02\x00J\x00\x00\x00\xd6\x0e\x00\x00\x96\x00\x02\x00\x10\x00\x00\x00 \x0f\x00\x00\x97\x00\x07\x00\x00\x04\x00\x000\x0f\x00\x00\x98\x00\x03\x00\x04\x00\x00\x000\x13\x00\x00\x99\x00\x04\x00}\x00\x00\x008\x13\x00\x00\x9a\x00\x04\x00\x05\x00\x00\x00,\x15\x00\x00\xa0\x00\x03\x00\x10\x00\x00\x00@\x15\x00\x00\xaa\x00\x03\x00\x06\x00\x00\x00`\x15\x00\x00\xb4\x00\x03\x00\x01\x00\x00\x00\x01\x00\x00\x00\xd0\x00\x04\x00\x01\x00\x00\x00\x00\x00\x00\x00\xe0\x00\x03\x00\x11\x00\x00\x00l\x15\x00\x00\x01@\x03\x008\x06\x00\x00\x8e\x15\x00\x00\x08@\x03\x00\x03\x00\x00\x00\xfe!\x00\x00\t@\x03\x00\x03\x00\x00\x00\x04"\x00\x00\x10@\x02\x00 \x00\x00\x00\n"\x00\x00\x11@\x07\x00\xfc\x00\x00\x00*"\x00\x00\x12@\x02\x00 \x00\x00\x00&#\x00\x00\x13@\x04\x00\x0b\x00\x00\x00F#\x00\x00\x15@\x07\x00d\x02\x00\x00r#\x00\x00\x16@\x04\x00\n\x00\x00\x00\xd6%\x00\x00\x18@\x04\x00\r\x00\x00\x00\xfe%\x00\x00\x19@\x07\x00\x1e\x00\x00\x002&\x00\x00 @\x04\x00\x07\x00\x00\x00P&\x00\x00!@\x04\x00\x05\x00\x00\x00l&\x00\x00$@\x04\x005\x00\x00\x00\x80&\x00\x00%@\x04\x00\t\x00\x00\x00T\'\x00\x00\'@\x04\x00\x06\x00\x00\x00x\'\x00\x00\x00\x00\x00\x00b\x00\x02\x00\x00\x00\x03\x00\x00\x00\x00\x00\x00\x00\x03\x00\x00\x00\x07\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\xff\x7f\xff\x7f\x03\x00\x02\x00\x00\x00\x04\x00\xff\xff\xf8\x00#\x00\x12\x00\x01\x004\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\xff\xff\xff\xff\xff\xff\x00\x00\x00\x00\x00\x00\x00\x00\xff\xff\xff\xff\x00\x00\x00\x00\xff\x7f\xff\xff\xff\xff\x00\x00\x00\x00\x04\x00\x00\x00\x12\x00\xaec\x1dH\x00\x00\x00\x00\x00\x00\x00\x00D\x00\x00\x00\xa0\x00\x80\x00\x00\x01 \x01\x00\x00\x03\x00\x03\x00\x00\x00\x08\x00\x08\x00\x99\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x01 \x01i\x00\x00\x00\x00\x00\xf8\x00\xff\xff\xff\xff\xff\xff\xff\xff\x00\x00\x00\x00\x00\x00Canon EOS 80D\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00Firmware Version 1.0.1\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xaa\xaa\x80H\x80HH\x00\x01<\x01g\x00\x03\x00\x00\x00\x00\x00\x00\x01\x00\x00\x06\x00\x00\x00\x99\x00\x00=\x00=\x00=\x00\x12\x00\x00\x00\x00\x00\x00\x02\x00\x02\x03\x00\x00\x00\x00\x00\x17\x00\x12\x00\x00\x00\x00\x02\xbb\xbb\x16\x16\x17\x10\xff\xc2\xff\xff\x96\xff\x9a\x8fX\x00\x00}6\x00\x15\x15\x00\x01"\x11\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x8e\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xff\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\xff\xff \x00\x01\x01\x00\x00\x00\x00\x03\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xd5\x00\x93\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x7f\xff\x7f\xff\xff\x0c\xee\xee\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xaa\xbb/\x02h\x02\x8f\x02u\x00\x00\x00\x00\x00\x00\x00\x17\x00\x00\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\x00\x00\xff\xff\xff\xff\xff\xff\xff\xff\x00\x14\x1f\x00\x00\x00\x00\x00\x00\x00\xdd\xdd\x03\x00\x00\x00\x03\x00\x00\x00\x00\x00\x00\x00H\x00\x00\x00\x00\x00\x00\x00\x00}\x06\x00\x00\x00\x17\x00\x03\x00\x00\x00P\x14\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x07\x00\x00\x00\x00\x00\x00\x00\x03\x00\x00\x00\x03\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x87\x00\x00\x00\x01\x00\x00\x00\x01\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x15H\x00\xf8\x00\x12\x00#\x91\x05\x00\x00\x00\xff\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x15\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x03\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\n\x02\x00\x01\x00\x00\x00\x00\x00\x02\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x1c\x0e\x07\x03\x81\xc0\x00\x00\x00\xff\xff\xff\xff\xff\xf8\x00\x00\x00\x10\x00\x00\x00\x00\x00\x01\x00\x00\x00\xff\xff\xff\xff\x7f\xff\x00\x00\t\x00\x00\x00\x10\x00\x00\x00\x10\x00\x00\x00\x00\t\x00\x00\x00\x12\x00\x00\xf8\x00\x12\x00#\x91\x05\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x15\x15H\xa3\x02\x16==\x00==\x00\x12\x00\x12\x00\xd5\x00\x93\x00\x00\x00\x80\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xc4\t\x00\x00\x0c\x00\x00\x00\x00\x00\x00\x00\x03\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00@\x14\x00\x00\x80\r\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xd0\x02\x00\x00\xe0\x01\x00\x00@\x14\x00\x00\x80\r\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xd0\x02\x00\x00\xe0\x01\x00\x00\xd0\x02\x00\x00\xe0\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xd0\x02\x00\x00\xe0\x01\x00\x00\xd0\x02\x00\x00\xe0\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xd0\x02\x00\x00\xe0\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xff\xff\xff\xff\x07\x00\x00\x00\x00\x00\x00\x00\xff\xff\xff\xff\x07\x00\x00\x00\x00\x00\x00\x00\x01\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x03,\x01\x00\x00\xb5\x0b\x00\x00\x00\x00\x00\x03,\x01\x00\x00\xb5\x0b\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x04\x00\x04\x00\x04\x00\x04\x00\x00\x00\x00\xf8\xddf@\x00\x00\x00\x001.0.1\x009D(84)\x00\x00\x00\x00\x0c\x9b\x00\x80\xff\xff\xff\xff\x00\x00\x00\x00\x13\x00g\x00\xe4j\x89\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x8e\x00\x00\x00d\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00s\x1f\x00\x00\x00\x00\x00\x00d\x00\x00\x00e\x00\x00\x00\x00\x00\x00\x00\x08\x00\x00\x00\x08\x00\x00\x00\x08\x00\x00\x00\x08\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x006\x02\x00\x04\x00\x04d\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x006\x02\x00\x04\x00\x04d\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x006\x02\x00\x04\x00\x04d\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x006\x02\x00\x04\x00\x04d\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x006\x02\x00\x04\x00\x04d\x02\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x03\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xef\xbe\xad\xde\xef\xbe\xad\xde\x04\x00\x00\x00\x04\x00\x00\x00\x00\x00\x00\x00\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xef\xbe\xad\xde\xef\xbe\x00\x00\x9f\x00\x07\x00p\x00\x8c\x01\x00\x00-\x00-\x00p\x17\xa0\x0fp\x17\xa0\x0f\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\xfa\x00\x9e\x05A\x04\xf1\x01\x00\x00\x0f\xfe\xbf\xfbb\xfa\x04\xf9A\x04\xf1\x01\x00\x00\x0f\xfe\xbf\xfbb\xfa\x04\xf9\xfc\x06\xf1\x01\x00\x00\x0f\xfe\xbf\xfbb\xfa\x04\xf9\xfc\x06\x9e\x05\x00\x00\x0f\xfe\xbf\xfbb\xfa\x04\xf9\xfc\x06\x9e\x05A\x04\x0f\xfe\xbf\xfbb\xfa\x04\xf9\xfc\x06\x9e\x05A\x04\xf1\x01\xfc\x06\x9e\x05A\x04\xf1\x01\x00\x00\xbb\x02\xbb\x02\\\x03\\\x03\\\x03\xbb\x02\xbb\x02\xbb\x02]\x01\xae\x01\xae\x01\xae\x01]\x01]\x01]\x01\xbb\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00]\x01]\x01R\xfeR\xfe\xa3\xfe\xa3\xfe\xa3\xfe\x00\x00\x00\x00\x00\x00\xa4\xfcE\xfdE\xfdE\xfd\xa3\xfe\xa3\xfe\xa3\xfeR\xfeE\xfdE\xfdE\xfd\xa4\xfc\xa4\xfc\x00\x00\x00\x00\x00\x00\x1c\x0e\x07\x03\x81\x18\x00\x00\x00\x00\x00\x00\xff\xff\x10\x00\x00\x00<\x00\x00\x00\x13\x00\x00\x00\x00\x00\x00\x00L\x00\x00\x00LP-E6N\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00L\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xff\xff\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xff\xff\xff\xff=\x00\xff\xff\x00\x00\x00\x00\xd5\x00\x93\x00\x00\x00\x00\x00\x00\x00\x00\x00\xff\xff\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x004\x00\xff\x7f\xff\x7f\xff\xff\x00\x0018-35mm\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xf33xzUD0030168\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xf4\x01\x00\x00\x03\x00\x00\x00\x01\x00\x00\x00T\x00\x00\x00\x06\x00\x00\x00\x01\x01\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x02\x01\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x04\x01\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x05\x01\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x06\x01\x00\x00\x02\x00\x00\x00\x03\x00\x00\x00\x00\x00\x00\x00\x08\x01\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x03\x00\x00\x00\x0c\x01\x00\x00\x10\x00\x00\x00\x02\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x18\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x04\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x19\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x1a\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x0e\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x05\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x12\x05\x00\x00\x02\x00\x00\x00\x00\x00\x00\x00G\x00\x00\x00\x1b\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x16\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x1e\x05\x00\x00\x01\x00\x00\x00\x02\x00\x00\x00\x1c\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x13\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x0c\x05\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x1d\x05\x00\x00\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x07\x05\x00\x00\x10\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x04\x00\x00\x00\x80\x00\x00\x00\x04\x00\x00\x00\n\x04\x00\x00\x01\x00\x00\x00\xc3\x00\x00\x00\x06\x07\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x14\x08\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x0c\x07\x00\x00\x13\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00p\x17\x00\x00\xa0\x0f\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00 \x00\x00\x00\x03\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xff\xffP\x14\x87\x00\x00\x00\x00\x00\x00\x00\x04\x00\x04\x00\x0c\x00>\x02\x00\x04\x00\x04,\x04\x01\x00"\x00\x90\x18\xd8\x0f\x01\x00\x01\x00\x14\x01.\x00\x83\x18\xcd\x0f\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\r\x00\x95\x03\x00\x04\x00\x04}\x01x\x02\x00\x04\x00\x04\xfd\x01\xbc\x01\x00\x04\x00\x04\xb5\x02N\x06\x1c\x07\x1e\x07\x9f\x02\xf3\x06X\x0b_\x0b\x97\x05/\x03W\x07^\x07\xf4\x04\x00\x00\x00\x00\xfd\x00\x02\x01\xf9\x00\x00\x00\x9f\x05L\nS\n\xd6\x05l\x02\xb6\x00\xb6\x00%\x00\x0c\x05*\x07+\x07x\x01\x88\x00o\x02t\x02\xfd\x02\xa6\x05}\n\x88\n\xff\x05s\x02\xc1\x00\xc2\x00%\x00$\x05k\x07n\x07\x80\x01\x96\x00\x98\x02\x9d\x02!\x03%\x05\x00\x04\x00\x04\xc1\t\x84\x0c>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c\x00\x04\x00\x04\x00\x04\x00\x04\x1b\x13\x00\x04\x00\x04\x00\x04\x00\x04\x1b\x13>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c\x00\x0e\xed\x03\x11\x04L\x03\x94*\x00\x0e\xed\x03\x11\x04L\x03\x94*A\x02\x92\x04\x92\x046\x03E\x1bF\x07\x00\x04\x00\x04\xba\x06P\x14X\x08\x00\x04\x00\x04\xd1\x05X\x1b\xc6\x07\x00\x04\x00\x04C\x06p\x17%\x05\x00\x04\x00\x04\xc1\t\x80\x0c\x18\x06\x00\x04\x00\x04*\th\x0eF\x07\x00\x04\x00\x04\xba\x06P\x14\xfc\x07\x00\x04\x00\x04.\x06s\x18F\x07\x00\x04\x00\x04\xba\x06P\x14F\x07\x00\x04\x00\x04\xba\x06P\x14F\x07\x00\x04\x00\x04\xba\x06P\x14F\x07\x00\x04\x00\x04\xba\x06P\x14F\x07\x00\x04\x00\x04\xba\x06P\x14\xfb\x03\x00\x04\x00\x04\xf1\x03P\x13\xfb\x03\x00\x04\x00\x04\xf1\x03P\x13\xfb\x03\x00\x04\x00\x04\xf1\x03P\x13\xfb\x03\x00\x04\x00\x04\xf1\x03P\x13\xfb\x03\x00\x04\x00\x04\xf1\x03P\x13>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c\x00\x04\x00\x04\x00\x04\x00\x04\x1b\x13\x00\x04\x00\x04\x00\x04\x00\x04\x1b\x13>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c>\x08\x00\x04\x00\x04\x95\x05\xcf\x1c\x00\x0e\xed\x03\x11\x04L\x03\x94*\r\xff\xae\x015\x03\x94*!\xff\xb8\x01\x1f\x03\x10\'L\xff\xd0\x01\xef\x02l y\xff\xeb\x01\xc0\x02X\x1b\xaf\xff\x0f\x02\x8e\x02p\x17\xc7\xff \x02y\x02\xe0\x15\xe3\xff3\x02a\x02P\x14\x12\x00T\x027\x02\\\x12L\x00\x82\x02\x0c\x02h\x10~\x00\xaf\x02\xe8\x01\xd8\x0e\xb1\x00\xdf\x02\xc8\x01\xac\r\xed\x00\x1c\x03\xa4\x01\x80\x0c\x1b\x01M\x03\x89\x01\xb8\x0bG\x01\x88\x03t\x01\xf0\n\xb1\x01!\x04J\x01`\t\xf4\x01\x0e\x08\x1e\x08\x00\x02\x00\x02\x00\x02\x00\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x0c\x00\r\x00\x0b\x00\x0c\x00\x0b\x00\t\x00\x00\x00\x00\x00\x00\x00\x0c\x00\x0b\x00\n\x00\x0b\x00\x0b\x00\t\x00\x00\x00\x00\x00\x00\x00\x0c\x00\x0b\x00\x0b\x00\x0b\x00\n\x00\t\x00\x07\x00\x00\x00\x00\x00\x0b\x00\x0b\x00\x0b\x00\n\x00\n\x00\x08\x00\x01\x00\x00\x00\x00\x00\x0b\x00\x0b\x00\x0b\x00\r\x00\x0c\x00\x0b\x00\x05\x00\x00\x00\x00\x00\t\x00\r\x00\n\x00\x0f\x00\x0b\x00\t\x00\x04\x00\x00\x00\x00\x00\x06\x00\x0e\x00\n\x00\x08\x00\x04\x00\x00\x00\x00\x00\x00\x00\x00\x00U\x00~\x00y\x00~\x00s\x00m\x00O\x00I\x00J\x00b\x00\x82\x00{\x00\x81\x00{\x00x\x00O\x00I\x00H\x00o\x00\x81\x00|\x00\x7f\x00|\x00z\x00T\x00J\x00F\x00x\x00|\x00\x80\x00|\x00|\x00y\x00\\\x00K\x00G\x00z\x00x\x00\x82\x00z\x00z\x00z\x00g\x00K\x00E\x00y\x00w\x00\x86\x00z\x00y\x00x\x00\\\x00J\x00B\x00s\x00s\x00\x82\x00q\x00g\x00Y\x00M\x00J\x00B\x00\x06\x00d\x00P\x00]\x00d\x00<\x00\x00\x00\x00\x00\x00\x00$\x00d\x00d\x00d\x00d\x00S\x00\x00\x00\x00\x00\x00\x00<\x00d\x00d\x00d\x00d\x00d\x00\x06\x00\x00\x00\x00\x00]\x00d\x00d\x00`\x00d\x00d\x00\x17\x00\x00\x00\x00\x00d\x00d\x00d\x00d\x00d\x00d\x002\x00\x00\x00\x00\x00d\x00L\x00d\x00S\x00Z\x00P\x00\n\x00\x00\x00\x00\x00]\x002\x00`\x00$\x00\x14\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x80\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xf2\x0fa\x00@\x00m\x10\x0f\x10\xa1\xff\xc2\xff\x97\x0f\x00\x00l\xa2\x01\x00@i\x05\x00\x80\xc4\x06\x00\xfch\x06\x00,\xa3\x00\x04\x00\x04\x00\x04\x00\x00\x00\x00\x00\x00\xfd\xff\xff\x1f\x00\x01\x00\x00\x00\x00\x00\x04$\x03\xa8\x016\x02d\x02\xd5\x01\xe5\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x1f\x00?\x00_\x00\x7f\x00\x9f\x00\xbf\x00\xdf\x00\xff\x00\x00\x00\x1f\x00?\x00_\x00\x7f\x00\x9f\x00\xbf\x00\xdf\x00\xff\x00\x00\x00\x1f\x00?\x00_\x00\x7f\x00\x9f\x00\xbf\x00\xdf\x00\xff\x00\x00\x00\x1f\x00?\x00_\x00\x7f\x00\x9f\x00\xbf\x00\xdf\x00\xff\x00\x00\x00\x1f\x00?\x00_\x00\x7f\x00\x9f\x00\xbf\x00\xdf\x00\xff\x00\x00\x00\x1f\x00?\x00_\x00\x7f\x00\x9f\x00\xbf\x00\xdf\x00\xff\x00\x00\x00\x03\x00\x00\x00\x10\x00 \x00@\x00`\x00\x80\x00\xc0\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xe8\x03\xe8\x03\xe8\x03\xe8\x03\xe8\x03\xe8\x03\xe8\x03\x88\x04\x00\x00\x00\x02\x00\x02\xff\x01\xff\x01\xab,t.\x10\'\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x03\x00\xd2\x00\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x00\x03\x00\xd2\x00\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x00\x03\x00\xd2\x00\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x00\x03\x00\xd2\x00\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x00\x03\x00\xd2\x00\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x00\x03\x00\xd2\x00\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x03\x00\x03\x00\x03\x00\x03\x00\x03\x00\x03\x00\x03\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00B\x00B\x00B\x00B\x00B\x00B\x00\x02\x00\x11\x00\xff\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00M\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x88\x13\x87\x05\x00\x04b\x03\x00\x00\x00\x00\x00\x00\x14\x00\x00\x00\xea\x01\xaf\x00G\x04\n\x01)\x01(\x00\xe7\x04\xb8\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xf1\x01\x00\x04\x00\x04\xde\x02d\x00\x00\x00\x15\x00\x00\x00\xec\x01\xac\x00I\x04\r\x01d\x00i\x00H\x00\x80\x00H\x00\xf8\x00\xff\x00\xfd[\xbaf1\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x1f\x00?\x00_\x00\x7f\x00\x9f\x00\xbf\x00\xdf\x00\xff\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00d\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x1f\x00?\x00_\x00\x7f\x00\x9f\x00\xbf\x00\xdf\x00\xff\x00\xf3\x00\x8e\x00\r\x00\x03\x00\x02\x00\x02\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xff\x00\x0f\x00i\x00H\x00\xfd[Ff\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xf8\x00\xff\x00H\x00+\x01+\x00\xe6\x04\xbe\x00p\x01\x00\x00\x00\x00\x80\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xee\xff\xbf\xbf\xfb\xb7\xf9\xfd\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x1f\x00\x00\x00\x00\x00d\x00\x00\x00\x00\x00\x00\x00\x00\x00d\x00d\x00d\x00d\x00\x00\x00d\x00d\x00=\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00h\x00H\x00\xfd[5f\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xf8\x00\xff\x00H\x00\xf4\x01\xab\x00\x00\x00\x00\x00\xfd[\xbbf\x1e\x00i\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x1f\x00?\x00_\x00\x7f\x00\x9f\x00\xbf\x00\xdf\x00\xff\x00\x00\x00\x00\x00\xf1\x01\x00\x04\xde\x02\x00\x00\xff\xdf\xfe\xee;\xfb\xbf\xb3\xed\xfe\xfe\xcf3\xff\xff\xbf\xfe\xde\xed\xdf\xbf\xfb\xf7\xb7\xef\xdc\xdd\xff\xbf{~\xbb\xef\xff\xdd\xcdw\xbd\xff\xfd\xfd\xed\xff\x9f\xff\xfb;{\xff\xfc\xfd\xff\xaf\xff\xf7\xbb\xfd\xed\xdf\xeb\xbfs\xbb7\x87\x00\x87\x00\x87\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00,\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\n\x00\x00\x00\xff\xff\xff\xff\x00\x00\x00\x00\n\x00\x00\x00\x00\x00\x00\x00\n\x00\x00\x00\x00\x00\x00\x00\n\x00\x00\x00\x000d\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00F\x00\x00\x00d\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x88\x0ep\x17\xa0\x0f\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x12\x00\x00\x00\x00\x00\x00\x00(\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x004\x00\x00\x00\x00\x00\x00\x00\x03\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x1c\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xff\xff\xff\x7f\x00\x00\x00\x00\x01\x00\x00\x00\x14\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\xd4\x00\x00\x00\x07\x00\x00\x00\x01\x00\x00\x00\x14\x00\x00\x00\x01\x00\x00\x00\x01\x01\x00\x00\x01\x00\x00\x00\xff\xff\xff\xff\x02\x00\x00\x00\x14\x00\x00\x00\x01\x00\x00\x00\x01\x02\x00\x00\x01\x00\x00\x00\xff\xff\xff\xff\x03\x00\x00\x00\x14\x00\x00\x00\x01\x00\x00\x00\x01\x03\x00\x00\x01\x00\x00\x00\xff\xff\xff\xff\x04\x00\x00\x008\x00\x00\x00\x04\x00\x00\x00\x01\x04\x00\x00\x01\x00\x00\x00\xff\xff\xff\xff\x02\x04\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x03\x04\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x04\x04\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x05\x00\x00\x00\x14\x00\x00\x00\x01\x00\x00\x00\x01\x05\x00\x00\x01\x00\x00\x00\xff\xff\xff\xff\x06\x00\x00\x00\x14\x00\x00\x00\x01\x00\x00\x00\x01\x06\x00\x00\x01\x00\x00\x00\xff\xff\xff\xff\x07\x00\x00\x00\x14\x00\x00\x00\x01\x00\x00\x00\x01\x07\x00\x00\x01\x00\x00\x00\xff\xff\xff\xff$\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x18\x00\x00\x00\x02\x02\x06\x00\x84\x00\x00\x9d\x00\x00\x00\x0b\xd0\xd0\xd0\x00\x00\x00\x00\x00II*\x00\x8e\x03\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x02\x00\x01\x00\x02\x00\x04\x00\x00\x00R98\x00\x02\x00\x07\x00\x04\x00\x00\x000100\x00\x00\x00\x00\x80\x8d[\x00\x99\x03\x00\x00\x00\t=\x00R\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00123021001945\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x12\x00\x00\x00\x01\x00\x00\x00#\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00\x01\x00\x00\x0018-35mm\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x000000000000\x00\x00\x01\x00\x00\x00\x01\x00\x04\x00\x00\x00\x02\x03\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x06\x00\x03\x01\x03\x00\x01\x00\x00\x00\x06\x00\x00\x00\x1a\x01\x05\x00\x01\x00\x00\x00\xd40\x00\x00\x1b\x01\x05\x00\x01\x00\x00\x00\xdc0\x00\x00(\x01\x03\x00\x01\x00\x00\x00\x02\x00\x00\x00\x01\x02\x04\x00\x01\x00\x00\x00\xe40\x00\x00\x02\x02\x04\x00\x01\x00\x00\x00Y\x08\x00\x00\x00\x00\x00\x00H\x00\x00\x00\x01\x00\x00\x00H\x00\x00\x00\x01\x00\x00\x00\xff\xd8\xff\xdb\x00\x84\x00\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x02\x01\x01\x01\x01\x01\x02\x02\x02\x02\x02\x03\x03\x03\x03\x03\x03\x03\x03\x03\x04\x04\x04\x03\x03\x04\x03\x03\x03\x04\x05\x04\x04\x05\x05\x05\x05\x05\x03\x04\x05\x06\x05\x05\x06\x04\x05\x05\x05\x01\x01\x01\x01\x01\x01\x01\x02\x01\x01\x02\x05\x03\x03\x03\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\xff\xc0\x00\x11\x08\x00x\x00\xa0\x03\x01!\x00\x02\x11\x01\x03\x11\x01\xff\xc4\x01\xa2\x00\x00\x01\x05\x01\x01\x01\x01\x01\x01\x00\x00\x00\x00\x00\x00\x00\x00\x01\x02\x03\x04\x05\x06\x07\x08\t\n\x0b\x01\x00\x03\x01\x01\x01\x01\x01\x01\x01\x01\x01\x00\x00\x00\x00\x00\x00\x01\x02\x03\x04\x05\x06\x07\x08\t\n\x0b\x10\x00\x02\x01\x03\x03\x02\x04\x03\x05\x05\x04\x04\x00\x00\x01}\x01\x02\x03\x00\x04\x11\x05\x12!1A\x06\x13Qa\x07"q\x142\x81\x91\xa1\x08#B\xb1\xc1\x15R\xd1\xf0$3br\x82\t\n\x16\x17\x18\x19\x1a%&\'()*456789:CDEFGHIJSTUVWXYZcdefghijstuvwxyz\x83\x84\x85\x86\x87\x88\x89\x8a\x92\x93\x94\x95\x96\x97\x98\x99\x9a\xa2\xa3\xa4\xa5\xa6\xa7\xa8\xa9\xaa\xb2\xb3\xb4\xb5\xb6\xb7\xb8\xb9\xba\xc2\xc3\xc4\xc5\xc6\xc7\xc8\xc9\xca\xd2\xd3\xd4\xd5\xd6\xd7\xd8\xd9\xda\xe1\xe2\xe3\xe4\xe5\xe6\xe7\xe8\xe9\xea\xf1\xf2\xf3\xf4\xf5\xf6\xf7\xf8\xf9\xfa\x11\x00\x02\x01\x02\x04\x04\x03\x04\x07\x05\x04\x04\x00\x01\x02w\x00\x01\x02\x03\x11\x04\x05!1\x06\x12AQ\x07aq\x13"2\x81\x08\x14B\x91\xa1\xb1\xc1\t#3R\xf0\x15br\xd1\n\x16$4\xe1%\xf1\x17\x18\x19\x1a&\'()*56789:CDEFGHIJSTUVWXYZcdefghijstuvwxyz\x82\x83\x84\x85\x86\x87\x88\x89\x8a\x92\x93\x94\x95\x96\x97\x98\x99\x9a\xa2\xa3\xa4\xa5\xa6\xa7\xa8\xa9\xaa\xb2\xb3\xb4\xb5\xb6\xb7\xb8\xb9\xba\xc2\xc3\xc4\xc5\xc6\xc7\xc8\xc9\xca\xd2\xd3\xd4\xd5\xd6\xd7\xd8\xd9\xda\xe2\xe3\xe4\xe5\xe6\xe7\xe8\xe9\xea\xf2\xf3\xf4\xf5\xf6\xf7\xf8\xf9\xfa\xff\xda\x00\x0c\x03\x01\x00\x02\x11\x03\x11\x00?\x00\xff\x00?\xfa(\x00\xa2\x80\n(\x00\xa2\x80\n(\x00\xa2\x80\n(\x00\xa2\x80\n(\x00\xa2\x80\n(\x00\xa2\x80\n(\x00\xa2\x80\n(\x00\xa2\x80\n(\x00\xa2\x80\n(\x00\xa2\x80\n(\x00\xa2\x80\n\xbfe\x84\xdf)\x19,\x1a\x08\xc9\xc0\xc1 \xee<\xf1\xf7x\xe7\xfb\xd9\xea+l?\xf1\xa2\xfb\x7f_\x98\xa5\xb3(QX\x8cr\x82\xcc\xaa\xa3s3\x00\x17\xd4\xd3\xe5WI\n8\xdaW\x8cd\x1c\x0f\xc3\xf3\xabI\xf29X:\x91QP\x01E\x00\x14P\x01E\x00\x14P\x01E\x00\x14P\x01Z#\t4\x10\x021\x0cd\xb9\x1cfF\x04\xb7\\\xf4\xe1}>\\\xf7\xad\xa8\xe9$\xfc\xd7\xe6&gQX\x8c\xb1nv\xb3K\xde%%}\x98\xf0\x0f\xd4\x13\x9f\xc0S\xee\xb9m\xfdHw\x89\x89\xeapx\xfd\x08\x1f\x87\xbdo\xff\x00.m\xf3\x15\xb5\xb9R\xa7h\xd4[\xc7(\xce\xe3##zc\xb7\xf5\xfc\xab8\xa4\xf9\xaf\xd8d\x14T\x00Q@\x05\x14\x00Q@\x05\x14\x00Q@\x16-\x95L\xa1\xa4\xc7\x97\x103H\x0f\xf1\x01\xce?\x13\x81\xf8\xfbS\xa0v{\x90\xec~g,\xc4\x8c\x0eH5\xbd=\x1c<\xdf\xf5\xfa\x89\xf5*\xd1X\x0c\xb0~H\xa3^wJ\xfekzm\x1c/\xeb\xbb\xf4\xa7\xc8\x035\xd2\x13\x82\xaceL\xf7 \xf2?#\x9f\xf8\x0f\xbdo\xd1\xafO\xc9\x8b\xa9R\xad\xaf0\x04\xc6w\xf9\x9b}\x88\xda\x7f\xc4~\'\xd6\xa2\x9f\xda\xf4\xfdP\xca\x94V`\x14P\x01E\x00\x14P\x01E\x00\x14P\x05\xaf\xf5v\xa4\xff\x00\x1d\xc3\xe0\x0ctE\xef\xf8\xb1\xc7\xfc\x00\xd3mN&Nq\xc3s\x90;\x1a\xde?\xc4\xa6\xbd?;\x8b\xb9^\x95T\xb3\x05Q\x96b\x14\x0fSX\xa5v\x92\x19,\xcc\xa6R\x10\xe5#\xc4h\xc0c q\x9c{\xf5\xc7\xbdK!U\xbc|\x9c)\x90\xab\x1f@x?\xce\xb6r\xf8\x9a\xee\xbfQv*\xb2\x94fV\x18e%X\x1e\xc4U\x85m\x91\xc0\xc7\xa0\x95\xf3\x8e\xa4|\xb9\xfeu0\xf7\\\x9fo\xf3A\xd8\x86D1\xbb\xa19(\xc5r:\x1aef\xd5\x9bC\n)\x00Q@\x05\x14\x00Q@\x05*\xa9b\x15Afb\x15UA$\x9a\x12\xbe\x88\t\xee\x8a\x89|\xb4ee\x81D!\xd3\xee\xb1\x1dH\xf6\'$\x1fLQjq:v\xe1\xbd\xbb\x1a\xda6\xf6\xe9.\xe8KU\xa9^\xa6\x84`\xbc\x9d\xa2\x8c\xbf\xe3\xd0~\xa4TC\xe2C!\xa9\xeey\xb8\x94\xff\x00\xb6h_\xc3~\xab\xf5\x0e\xa2Lw2\xberd@\xcc}OC\xf9\x91\x9f\xc6\x86\xff\x00Q\x17\xfdt\x93\xeb\xd1j\xbf\x9f\xd3\xf5@$\x9f\xf2\xcd\xbf\xbd\x18\xfc\xc7\x1f\xd2\xa2\xa8\x96\xed\x80QR\x01E\x00\x14P\x01E\x00\x15f\xd8a\xdeb\x06 \x8c\xcb\xce\x0f=\x17\xff\x00\x1e#\xf5=\xaa\xe9\xfch\x1e\xa5j\xb3g\xff\x00\x1f\x11\x81\x9c\x92@\xc62N\x0f\xaf\x14Sv\xa9\x17\xe6\x05j\x98\xfc\xb0\x01\xff\x00=d\xdd\xf8(\xc0\xfdI\xfc\xbd\xa9\xa5\xefI\xf6\xbf\xf9~\xa0CV.\xbf\xe3\xe2o\xf7\xcd%\xfc7\xea\xbfP\x1b\x9d\xd0\xe3\xbcO\x91\xcf\xf0\x9f\xf0?\xfa\x11\xa1\xbf\xd4E\xff\x00]$\xfeKT\xbe\x19?/\xd4]D\xce\xe8J\xf1\xfb\xb7\xde=px?\xc8~\xb5\x15D\xba\x0c(\xa9\x00\xa2\x80\n(\x00\xa2\x80\n\xb2\xc0\xc7k\x18 \x83;\x99\xbb\xf2\xa3*\x0f\xa7]\xe3\xf0\xab\x87\xda}\x97\xfc\x0f\xd4L\xadVl\xff\x00\xe3\xee\xd8c9\x9d\x060\x0ey\xa5\x1f\x89\x0c\xadSLF\xfd\x8b\xf7cQ\x18\xfc:\xfed\x93\xf8\xd5\xcbG?P!\xab\x17\x7f\xf1\xf37\xfdt5+\xf8r\xf5_\xa8\xba\xa1\xb0\xf2\xc6?\xf9\xe8\xa5\x07\xbbv\xfd@\x1f\x8d+\xff\x00\xc7\xbc=\x7f\xd6I\xdb\x8e\x8bW\x1f\xe1\xcb\xd3\xf5A\xd8l\\\xb3\'\xfc\xf4B\xa3\xdd\xba\x8f\xcc\x8c~5\x15f\xf6LaEH\x05\x14\x00Q@\x05\x14\x00\xe4F\x91\xd24\x05\x9d\xd8"\xa8\x04\x92O\x02\xa5\xb8`\xf3>\xd2\n.#B0AU\x18\x1d\x86x\x1d{\xfe5\x7fa\x8b\xa9\x05Y\xb3\x19\xbb\xb5\x1csq\x18\xe4\x12>\xf5J\xf8\x90\xc6@\x07\x98\x18\x80V0e \xf48\xe8?\x13\x81\xf8\xd4D\x92I\'$\x92I\xf55\xa5N\xaa\xddX\tV.\xbf\xe3\xe6o\xfa\xe8jW\xf0\xe4\xfc\xd7\xea.\xa8\x80\x12\x08#\xa8 \x8f\xad[\xb9\x18\x8e1\xd8\xc9#\xaf\xd0\x85"\xae\x1f\xc3\x9f\xf5\xd4\x1e\xe8\xa8\xa4\xab\x06\x1c\x15 \x82:\x83O\x95v\xc8\xe3\xa0\xceG\xa6:\x8f\xd2\xb3\xe8\xc6GEH\x05\x14\x00Q@\x05\x14\x01f\xdf\xe4\xf3f\xc0>\\Gnq\xc3\x1f\x94\x1f\xa8\xceA\x1d\xc0\xaa\xd5oH$\x1dB\xac\xd9\x7f\xc7\xe5\xa7_\xf8\xf9\x8f\xa6\x01\xfb\xc2\xa5|H\x06\xed\xd9\x0b\xb1\x1c\xbc\x9eR\xfe\x1c\xb7\xf3Z\x82\xb4\xad\xa4\xda\xfe\xb7\x00\xab\x17\x7f\xf1\xf3?\xfdt4\x97\xf0\xa5\xea\xbfQ}\xa4W\xabR\xb6\xeb[Pq\xf2<\xa889\xc7\xca\x7f\xad:o\xdd\x9a\xf2\xfdP=\xd1V\x8a\xc8aE\x00\x14P\x01E\x00\x14P\x05\xa9\x0f\x97m\x14X!\xa5cq!\xe4dr\xaa1\xd3\x8f\x98\xe4\x7f\x7f\xda\xaa\xd6\x957Q\xec\xbf\xaf\xc6\xe2ANV*\xca\xc3\xaa\xb0a\xf5\xa8\x8f\xc4\x86]\xd4\xbc\xb1{p\x91\x15(\x93:\xa9@@\xc6\xe2x\xfc\xf1\x8a\xa1ZVi\xd5\x93]\xc0*k\x87Y\'\x95\xd3%Y\xc9\x04\x8c\x12*S\\\x8e>k\xf5\x17T\xc8h\xa8\x18Q@\x05\x14\x00Q@\x05\x14\x00S\xe3F\x96D\x8d~\xf3\xb0Q\x9e\x99\xa6\x95\xda@>\xe1\xd1\xe6s\x16|\xa5\xf9!\xdc\xa0\x12\x83\x80H\xc9\x19 d\xe3\xb95\r9\xbb\xcd\xb0\n)\'f\x9b\x02\xcd\xec\xcbsywp\x9b\xb6\xcfs$\xcb\xbc\x00\xd8f\'\x9023\xcfj\xadNo\x9ar\x92\xea\x01EH\x05\x14\x00Q@\x05\x14\x00Q@\x05\x14\x00R\x82A\xc8$\x1fPph\x01(\xa0\x02\x8a\x00(\xa0\x02\x8a\x00(\xa0\x02\x8a\x00(\xa0\x02\x8a\x00(\xa0\x02\x8a\x00(\xa0\x02\x8a\x00(\xa0\x02\x8a\x00(\xa0\x02\x8a\x00(\xa0\x02\x8a\x00(\xa0\x0f\xff\xd9\x00\xff\xd9')
+
+    @keyword
+    def create_and_fill_config(self):
+        conf = ConfigData(False, '.', 10, 1, False, 1, 'test_img_0.png', None, None, 0, None, 'test_img_{}.png',
+                          0, 0, 0, 0)
+        conf.set('analyse_positions', '   line_edge_indices', '0 2')
+        conf.set('DEFAULT', '   date', '2018:11:27')
+        conf.save()
+
+    @keyword
+    def execute_ledsa_s1(self, use_config):
+        if use_config:
+            out = self.execute_ledsa('-s1')
+        else:
+            self.execute_ledsa('--config')
+            inp = b'test_img_0.png\ntest_img_0.png\n12:00:00\ntest_img_{}.png\n0\n0\n1\n'
+            out = self.execute_ledsa('-s1', inp)
+            check_error_msg(out)
+        return out[0].decode('ascii')[-7]
+
+    @keyword
+    def execute_ledsa(self, arg, inp=None):
+        p = Popen(['python', '-m', 'ledsa', arg], stdin=PIPE, stdout=PIPE, stderr=PIPE)
+        out = wait_for_process_to_finish(p, inp)
+        return out
+
+
+def create_img_array():
+    img = np.zeros((200, 50, 3), np.int8)
+    add_led(img, 50, 25)
+    add_led(img, 100, 25)
+    add_led(img, 150, 25)
+    return img
+
+
+def add_led(img, x_pos, y_pos):
+    rv = norm()
+    size = 20
+    led = np.zeros((size, size))
+    for x in range(size):
+        for y in range(size):
+            led[x, y] = calc_color_val(x, y, size, rv)
+    img[x_pos - size//2:x_pos + size//2, y_pos - size//2:y_pos + size//2, 0] = led
+    img[:, :, 1] = img[:, :, 0]
+    img[:, :, 2] = img[:, :, 0]
+
+
+def calc_color_val(x, y, size, rv):
+    dist = ((size/2 - x)**2 + (size/2 - y)**2)**0.5
+    scale = 1.7
+    return rv.pdf(dist / scale) * 350 * scale
+
+
+def wait_for_process_to_finish(p, inp=None):
+    out = p.communicate(inp)
+    check_error_msg(out)
+    return out
+
+
+def check_error_msg(out):
+    if out[1] is not None:
+        if out[1].decode('ascii') != '':
+            BuiltIn().log(out[1].decode('ascii'), 'ERROR')
+            exit()
