+++ ledsa/tests/AcceptanceTests/__main__.py
@@ -1,0 +1,6 @@
+from subprocess import call
+import pathlib
+
+
+path = pathlib.Path(__file__).parent.absolute()
+call(['python', '-m', 'robot', path])
