+++ ledsa/tests/plot_coordinates.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 from mpl_toolkits.mplot3d import Axes3D
 import numpy as np
-from ..core._led_helper import load_file
+from ..core.led_helper import load_file
 from os import sep
